% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_sf_backend.R
\name{as_sf_backend}
\alias{as_sf_backend}
\alias{as_sf_backend.SpatRaster}
\alias{as_sf_backend.RasterBrick}
\alias{as_sf_backend.DataBackendSpatRaster}
\alias{as_sf_backend.DataBackendRasterBrick}
\title{Convert to sf backend}
\usage{
as_sf_backend(data, primary_key = NULL, keep_rownames = FALSE, ...)

\method{as_sf_backend}{SpatRaster}(data, primary_key = NULL, keep_rownames = FALSE, ...)

\method{as_sf_backend}{RasterBrick}(data, primary_key = NULL, keep_rownames = FALSE, ...)

\method{as_sf_backend}{DataBackendSpatRaster}(
  data,
  primary_key = NULL,
  keep_rownames = FALSE,
  polygons = FALSE,
  ...
)

\method{as_sf_backend}{DataBackendRasterBrick}(
  data,
  primary_key = NULL,
  keep_rownames = FALSE,
  polygons = FALSE,
  ...
)
}
\arguments{
\item{data}{(\code{\link[=data.frame]{data.frame()}})\cr
The input \code{\link[=data.frame]{data.frame()}}.
Converted to a \code{\link[data.table:data.table]{data.table::data.table()}} automatically.}

\item{primary_key}{(\code{character(1)} | \code{integer()})\cr
Name of the primary key column, or integer vector of row ids.}

\item{keep_rownames}{(\code{logical(1)} | \code{character(1)})\cr
If \code{TRUE} or a single string, keeps the row names of \code{data} as a new column.
The column is named like the provided string, defaulting to \code{"..rownames"} for \code{keep_rownames == TRUE}.
Note that the created column will be used as a regular feature by the task unless you manually change the column role.
Also see \code{\link[data.table:as.data.table]{data.table::as.data.table()}}.}

\item{...}{(\code{any})\cr
Additional arguments passed to the respective \link[mlr3]{DataBackend} method.}

\item{polygons}{\verb{[logical]}\cr
Whether to convert to polygons instead of points.}
}
\value{
\link{DataBackend}.
}
\description{
Converts to a \link{DataBackendSF}.
}
