% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataBackendSpatRaster.R
\name{as_data_backend.SpatRaster}
\alias{as_data_backend.SpatRaster}
\title{Create a Data Backend}
\usage{
\method{as_data_backend}{SpatRaster}(data, primary_key = NULL, ...)
}
\arguments{
\item{data}{(\code{SpatRaster})\cr
A \code{SpatRaster} object to create a \link{DataBackend} from.}
}
\value{
\link{DataBackend}.
}
\description{
Wraps a \link{DataBackend} around \code{SpatRaster} objects.
}
