% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_task_classif.R
\name{as_task_classif.sf}
\alias{as_task_classif.sf}
\alias{as_task_classif.SpatRaster}
\title{Convert to a Classification Task}
\usage{
\method{as_task_classif}{sf}(
  x,
  target = NULL,
  id = deparse(substitute(x)),
  positive = NULL,
  label = NA_character_,
  task_train = NULL,
  ...
)

\method{as_task_classif}{SpatRaster}(
  x,
  target = NULL,
  id = deparse(substitute(x)),
  positive = NULL,
  label = NA_character_,
  task_train = NULL,
  ...
)
}
\arguments{
\item{x}{(any)\cr
Object to convert.}

\item{target}{(\code{character(1)})\cr
Name of the target column.}

\item{id}{(\code{character(1)})\cr
Id for the new task.
Defaults to the (deparsed and substituted) name of the data argument.}

\item{positive}{(\code{character(1)})\cr
Level of the positive class. See \link{TaskClassif}.}

\item{label}{(\code{character(1)})\cr
Label for the new instance.}

\item{task_train}{\verb{[mlr3::Task]}\cr
Task used for training the learner.}

\item{...}{(any)\cr
Additional arguments.}
}
\value{
\link{TaskClassif}.
}
\description{
Convert object to a \link[mlr3:TaskClassif]{mlr3::TaskClassif}.
}
