% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_task_regr.R
\name{as_task_regr.sf}
\alias{as_task_regr.sf}
\alias{as_task_regr.stars}
\alias{as_task_regr.SpatRaster}
\alias{as_task_regr.RasterBrick}
\alias{as_task_regr.RasterStack}
\title{Convert to a Regression Task}
\usage{
\method{as_task_regr}{sf}(
  x,
  target = NULL,
  id = deparse(substitute(x)),
  label = NA_character_,
  ...
)

\method{as_task_regr}{stars}(
  x,
  target = NULL,
  id = deparse(substitute(x)),
  label = NA_character_,
  ...
)

\method{as_task_regr}{SpatRaster}(
  x,
  target = NULL,
  id = deparse(substitute(x)),
  label = NA_character_,
  ...
)

\method{as_task_regr}{RasterBrick}(
  x,
  target = NULL,
  id = deparse(substitute(x)),
  label = NA_character_,
  ...
)

\method{as_task_regr}{RasterStack}(
  x,
  target = NULL,
  id = deparse(substitute(x)),
  label = NA_character_,
  ...
)
}
\arguments{
\item{x}{(any)\cr
Object to convert.}

\item{target}{(\code{character(1)})\cr
Name of the target column.}

\item{id}{(\code{character(1)})\cr
Id for the new task.
Defaults to the (deparsed and substituted) name of the data argument.}

\item{label}{(\code{character(1)})\cr
Label for the new instance.}

\item{...}{(any)\cr
Additional arguments.}
}
\value{
\link{TaskRegr}.
}
\description{
Convert object to a \link{TaskRegr}.
}
