% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{demo_raster}
\alias{demo_raster}
\title{Generate Demo Raster}
\usage{
demo_raster(dimension)
}
\arguments{
\item{dimension}{(\code{integer(1)})
xy dimension of raster}
}
\description{
Generates a square demo \link[terra:SpatRaster-class]{terra::SpatRaster}.
}
\keyword{internal}
