% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{demo_stack}
\alias{demo_stack}
\title{Generate a dummy 'terra::SpatRaster'}
\usage{
demo_stack(size = 50, layers = 5)
}
\arguments{
\item{size}{\verb{[integer(1)]}\cr
Size of raster stack in megabyte (disk space).}

\item{layers}{\verb{[integer(1)]}\cr
Number of layers.}
}
\description{
Generates a square demo \link[terra:SpatRaster-class]{terra::SpatRaster} stack.
}
