% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{demo_stack_rasterbrick}
\alias{demo_stack_rasterbrick}
\title{Generate a demo 'raster::brick'}
\usage{
demo_stack_rasterbrick(size = 50, layers = 5)
}
\arguments{
\item{size}{\verb{[integer(1)]}\cr
Size of raster stack in megabyte (disk space).}

\item{layers}{\verb{[integer(1)]}\cr
Number of layers.}
}
\description{
Generates a square demo \link[raster:brick]{raster::brick} object with options to set the size
(on disk) and number of layers.
}
\examples{
if (mlr3misc::require_namespaces("raster", quietly = TRUE)) {
  demo_stack_rasterbrick(size = 5, layers = 2)
}
}
