% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{demo_stack_spatraster}
\alias{demo_stack_spatraster}
\title{Generate a demo 'terra::SpatRaster'}
\usage{
demo_stack_spatraster(size = 50, layers = 5)
}
\arguments{
\item{size}{\verb{[integer(1)]}\cr
Size of raster stack in megabyte (disk space).}

\item{layers}{\verb{[integer(1)]}\cr
Number of layers.}
}
\description{
Generates a square demo \link[terra:SpatRaster-class]{terra::SpatRaster} object with options to set the
size (on disk) and number of layers.
}
\examples{
demo_stack_spatraster(size = 5, layers = 2)
}
