% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{generate_stack}
\alias{generate_stack}
\title{Raster Stack}
\usage{
generate_stack(layers, layer_size = NULL, dimension = NULL)
}
\arguments{
\item{layers}{List of \code{numeric_layer()} and \code{factor_layer()}\cr
List of layers.}

\item{layer_size}{\code{numeric(1)}\cr
Size of a single layer in Megabytes.}

\item{dimension}{\code{integer(1)}\cr
Number of rows and cols.

\code{layer_size} and \code{dimension} are mutually exclusive.}
}
\description{
Generates a square demo \link[terra:SpatRaster-class]{terra::SpatRaster} from a list of layers.
}
