% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LearnerClassifRangerParallel.R
\name{LearnerClassifRangerParallel}
\alias{LearnerClassifRangerParallel}
\title{Parallel Prediction Ranger Classification Learner}
\description{
Proof of concept for parallel prediction integrated into the \link[mlr3:Learner]{mlr3::Learner}
class. The Ranger classification learner is extended with the
\code{predict_newdata_parallel} method which splits \code{newdata} into \code{n} parts and
executes the prediction in parallel. \code{n} is equal to the number of available
cores.
}
\note{
The \code{predict_newdata_parallel} method is not useful for learning algorithms
with an integrated parallel predict function. For example, the \code{ranger}
learner predicts newdata internally on all available cores. Therefore, a new
learner property (\code{parallel_predict}) should indicate if
\code{predict_newdata_parallel} should be called at all.
}
\section{Super classes}{
\code{\link[mlr3:Learner]{mlr3::Learner}} -> \code{\link[mlr3:LearnerClassif]{mlr3::LearnerClassif}} -> \code{\link[mlr3learners:LearnerClassifRanger]{mlr3learners::LearnerClassifRanger}} -> \code{LearnerClassifRangerParallel}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{LearnerClassifRangerParallel$new()}}
\item \href{#method-predict_newdata_parallel}{\code{LearnerClassifRangerParallel$predict_newdata_parallel()}}
\item \href{#method-clone}{\code{LearnerClassifRangerParallel$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="base_learner">}\href{../../mlr3/html/Learner.html#method-base_learner}{\code{mlr3::Learner$base_learner()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="format">}\href{../../mlr3/html/Learner.html#method-format}{\code{mlr3::Learner$format()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="help">}\href{../../mlr3/html/Learner.html#method-help}{\code{mlr3::Learner$help()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="predict">}\href{../../mlr3/html/Learner.html#method-predict}{\code{mlr3::Learner$predict()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="predict_newdata">}\href{../../mlr3/html/Learner.html#method-predict_newdata}{\code{mlr3::Learner$predict_newdata()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="print">}\href{../../mlr3/html/Learner.html#method-print}{\code{mlr3::Learner$print()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="reset">}\href{../../mlr3/html/Learner.html#method-reset}{\code{mlr3::Learner$reset()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="train">}\href{../../mlr3/html/Learner.html#method-train}{\code{mlr3::Learner$train()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3learners" data-topic="LearnerClassifRanger" data-id="importance">}\href{../../mlr3learners/html/LearnerClassifRanger.html#method-importance}{\code{mlr3learners::LearnerClassifRanger$importance()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3learners" data-topic="LearnerClassifRanger" data-id="oob_error">}\href{../../mlr3learners/html/LearnerClassifRanger.html#method-oob_error}{\code{mlr3learners::LearnerClassifRanger$oob_error()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this \link[R6:R6Class]{R6} class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LearnerClassifRangerParallel$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-predict_newdata_parallel"></a>}}
\if{latex}{\out{\hypertarget{method-predict_newdata_parallel}{}}}
\subsection{Method \code{predict_newdata_parallel()}}{
Predicts \code{newdata} in chunks.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LearnerClassifRangerParallel$predict_newdata_parallel(newdata)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{newdata}}{(\code{data.frame()})\cr
New data to predict on}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LearnerClassifRangerParallel$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
