% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LearnerClassifSVMParallel.R
\name{LearnerClassifSVMParallel}
\alias{LearnerClassifSVMParallel}
\title{Parallel Prediction SVM Classification Learner}
\description{
Proof of concept for parallel prediction intergrated into the [mlr3::Learner]
class. The SVM classification learner is extended with the
`predict_newdata_parallel` method which splits `newdata` into `n` parts and
execudes the prediction in parallel. `n` is equal to the number of available
cores.
}
\note{
The `predict_newdata_parallel` method is not useful for learning algorithms
with an integrated parallel predict function. For example, the `ranger`
learner predicts newdata internally on all available cores. Therefore, a new
learner property (`parallel_predict`) should indicate if
`predict_newdata_parallel` should be called at all.
}
\section{Super classes}{
\code{\link[mlr3:Learner]{mlr3::Learner}} -> \code{\link[mlr3:LearnerClassif]{mlr3::LearnerClassif}} -> \code{\link[mlr3learners:LearnerClassifSVM]{mlr3learners::LearnerClassifSVM}} -> \code{LearnerClassifSVMParallel}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{LearnerClassifSVMParallel$new()}}
\item \href{#method-predict_newdata_parallel}{\code{LearnerClassifSVMParallel$predict_newdata_parallel()}}
\item \href{#method-clone}{\code{LearnerClassifSVMParallel$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="format">}\href{../../mlr3/html/Learner.html#method-format}{\code{mlr3::Learner$format()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="help">}\href{../../mlr3/html/Learner.html#method-help}{\code{mlr3::Learner$help()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="predict">}\href{../../mlr3/html/Learner.html#method-predict}{\code{mlr3::Learner$predict()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="predict_newdata">}\href{../../mlr3/html/Learner.html#method-predict_newdata}{\code{mlr3::Learner$predict_newdata()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="print">}\href{../../mlr3/html/Learner.html#method-print}{\code{mlr3::Learner$print()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="reset">}\href{../../mlr3/html/Learner.html#method-reset}{\code{mlr3::Learner$reset()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="train">}\href{../../mlr3/html/Learner.html#method-train}{\code{mlr3::Learner$train()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LearnerClassifSVMParallel$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-predict_newdata_parallel"></a>}}
\if{latex}{\out{\hypertarget{method-predict_newdata_parallel}{}}}
\subsection{Method \code{predict_newdata_parallel()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LearnerClassifSVMParallel$predict_newdata_parallel(newdata)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LearnerClassifSVMParallel$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
