% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TaskClassif_leipzig.R
\docType{data}
\name{leipzig}
\alias{leipzig}
\alias{mlr_tasks_leipzig}
\title{Leipzig Land Cover Task}
\source{
Copernicus Sentinel Data (2021). Retrieved from Copernicus Open Access Hub and processed by European Space Agency.
}
\description{
Point survey of land cover in Leipzig.
Includes Sentinel-2 spectral bands and NDVI.
}
\examples{
if(requireNamespace("sf")) {
library(sf)
data("leipzig", package = "mlr3spatial")
print(leipzig)
}
}
\keyword{data}
