% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{mask_stack}
\alias{mask_stack}
\title{Sample Points in Raster Stack}
\usage{
mask_stack(stack)
}
\arguments{
\item{stack}{(\link[terra:SpatRaster-class]{terra::SpatRaster})\cr
Raster stack.}

\item{n}{(\code{integer(1)})\cr
Number of points.}
}
\value{
\link[terra:SpatRaster-class]{terra::SpatRaster}
}
\description{
Masks stack to a circular area of interest.
}
\keyword{internal}
