% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{mlr3raster-package}
\alias{mlr3raster}
\alias{mlr3raster-package}
\title{mlr3raster: Parallel raster prediction for mlr3}
\description{
Parallel raster prediction for mlr3.
}
\references{
\cite{mlr3raster}{pkg::citation}
}
\author{
\strong{Maintainer}: Marc Becker \email{marcbecker@posteo.de}

}
