% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TaskClassif_leipzig.R
\docType{data}
\name{mlr_tasks_leipzig}
\alias{mlr_tasks_leipzig}
\title{Leipzig Land Cover Task}
\source{
Sentinel-2
}
\description{
Sentinel-2 spectral bands and NDVI to predict land cover in Leipzig.
}
\examples{
data("kc_housing", package = "mlr3data")
str(kc_housing)
}
\keyword{data}
