% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredictionRasterRegr.R
\name{PredictionRasterRegr}
\alias{PredictionRasterRegr}
\title{Demo title}
\description{
demo descrip
}
\section{Super class}{
\code{\link[mlr3spatial:PredictionRaster]{mlr3spatial::PredictionRaster}} -> \code{PredictionRasterRegr}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{PredictionRasterRegr$new()}}
\item \href{#method-clone}{\code{PredictionRasterRegr$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="mlr3spatial" data-topic="PredictionRaster" data-id="predict">}\href{../../mlr3spatial/html/PredictionRaster.html#method-predict}{\code{mlr3spatial::PredictionRaster$predict()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PredictionRasterRegr$new(stack, task, reclassify_table)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PredictionRasterRegr$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
