% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{sample_stack}
\alias{sample_stack}
\title{Sample Points in Raster Stack}
\usage{
sample_stack(stack, n = 100)
}
\arguments{
\item{stack}{(\link[terra:SpatRaster-class]{terra::SpatRaster})\cr
Raster stack.}

\item{n}{(\code{integer(1)})\cr
Number of points.}
}
\value{
\link[sf:sf]{sf::sf}
}
\description{
Samples \code{n} points of a raster stack.
}
\keyword{internal}
