% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{sample_vector}
\alias{sample_vector}
\title{Sample Stack}
\usage{
sample_vector(stack, n = 1000)
}
\arguments{
\item{stack}{\link[terra:SpatRaster-class]{terra::SpatRaster}\cr
List of layers.}

\item{n}{\code{numeric(1)}\cr
Number of sample points.}
}
\description{
Samples a \link[terra:SpatRaster-class]{terra::SpatRaster} and returns points.
}
