% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{weights_layer}
\alias{weights_layer}
\title{Weights Layer Generator}
\usage{
weights_layer(id, in_memory = FALSE)
}
\arguments{
\item{id}{(\code{character(1)})\cr
Layer id.}

\item{in_memory}{(\code{logical(1)})\cr
If \code{FALSE} (default), layer is written to disk.}
}
\description{
Generates a weights layer when passed to \code{\link[=generate_stack]{generate_stack()}}.
}
\keyword{internal}
