% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{write_raster}
\alias{write_raster}
\title{Write a Raster in Chunks}
\usage{
write_raster(data)
}
\arguments{
\item{data}{\verb{[SpatRaster]}\cr
\code{SpatRaster} object.}
}
\description{
Writes square raster to disk in chunks.
Internal helper function.
}
\keyword{internal}
