% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataBackendRaster.R, R/DataBackendVector.R
\name{as_data_backend.stars}
\alias{as_data_backend.stars}
\alias{as_data_backend.SpatRaster}
\alias{as_data_backend.RasterBrick}
\alias{as_data_backend.Raster}
\alias{as_data_backend.sf}
\title{Coerce to DataBackendRaster}
\usage{
\method{as_data_backend}{stars}(data, primary_key = NULL, train_task = NULL, ...)

\method{as_data_backend}{SpatRaster}(data, primary_key = NULL, train_task = NULL, ...)

\method{as_data_backend}{RasterBrick}(data, primary_key = NULL, train_task = NULL, ...)

\method{as_data_backend}{Raster}(data, primary_key = NULL, train_task = NULL, ...)

\method{as_data_backend}{sf}(data, primary_key = NULL, ...)
}
\arguments{
\item{data}{(\link[terra:SpatRaster-class]{terra::SpatRaster})\cr
The input \link[terra:SpatRaster-class]{terra::SpatRaster}.}

\item{primary_key}{(\code{character(1)} | \code{integer()})\cr
Name of the primary key column, or integer vector of row ids.}

\item{train_task}{\verb{[mlr3::Task]}\cr
Task used for training the learner.}

\item{...}{(\code{any})\cr
Not used.}
}
\value{
\link{DataBackend}.

\link{DataBackend}.
}
\description{
Wraps a \link{DataBackend} around spatial objects.
Currently this is only a synonym for \code{DataBackendRaster$new()} and does not
support coercing from other backends.

Wraps a \link{DataBackend} around spatial objects.
Currently this is only a synonym for \code{DataBackendVector$new()} and does not
support coercing from other backends.
}
