% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_rasterbrick_backend.R, R/as_sf_backend.R,
%   R/as_spatraster_backend.R, R/as_stars_backend.R
\name{as_rasterbrick_backend}
\alias{as_rasterbrick_backend}
\alias{as_sf_backend}
\alias{as_sf_backend.SpatRaster}
\alias{as_sf_backend.RasterBrick}
\alias{as_sf_backend.DataBackendSpatRaster}
\alias{as_sf_backend.DataBackendRasterBrick}
\alias{as_spatraster_backend}
\alias{as_stars_backend}
\alias{as_stars_backend.SpatRaster}
\alias{as_stars_backend.RasterBrick}
\alias{as_stars_backend.DataBackendSpatRaster}
\alias{as_stars_backend.DataBackendRasterBrick}
\title{Convert to and between spatial DataBackends}
\usage{
as_rasterbrick_backend(
  data,
  primary_key = NULL,
  response,
  response_is_factor = FALSE,
  ...
)

as_sf_backend(data, primary_key = NULL, ...)

\method{as_sf_backend}{SpatRaster}(data, primary_key = NULL, ...)

\method{as_sf_backend}{RasterBrick}(data, primary_key = NULL, ...)

\method{as_sf_backend}{DataBackendSpatRaster}(data, primary_key = NULL, polygons = FALSE, ...)

\method{as_sf_backend}{DataBackendRasterBrick}(data, primary_key = NULL, polygons = FALSE, ...)

as_spatraster_backend(data, primary_key = NULL, ...)

as_stars_backend(
  data,
  primary_key = NULL,
  response = NULL,
  response_is_factor = FALSE,
  quiet = FALSE,
  ...
)

\method{as_stars_backend}{SpatRaster}(
  data,
  primary_key = NULL,
  response = NULL,
  response_is_factor = FALSE,
  quiet = FALSE,
  ...
)

\method{as_stars_backend}{RasterBrick}(
  data,
  primary_key = NULL,
  response = NULL,
  response_is_factor = FALSE,
  quiet = FALSE,
  ...
)

\method{as_stars_backend}{DataBackendSpatRaster}(
  data,
  primary_key = NULL,
  response = NULL,
  response_is_factor = FALSE,
  quiet = FALSE,
  ...
)

\method{as_stars_backend}{DataBackendRasterBrick}(
  data,
  primary_key = NULL,
  response = NULL,
  response_is_factor = FALSE,
  quiet = FALSE,
  ...
)
}
\arguments{
\item{data}{(\code{\link[=data.frame]{data.frame()}})\cr
The input \code{\link[=data.frame]{data.frame()}}.
Converted to a \code{\link[data.table:data.table]{data.table::data.table()}} automatically.}

\item{primary_key}{(\code{character(1)} | \code{integer()})\cr
Name of the primary key column, or integer vector of row ids.}

\item{response}{(\code{\link{character}})\cr
The name of the response variable. Only needed when \code{response_is_factor = TRUE}.}

\item{response_is_factor}{(\code{\link{character}})\cr
When this backend should be used in a \link[mlr3:TaskClassif]{mlr3::TaskClassif}, set \code{response_is_factor = TRUE}.}

\item{...}{(\code{any})\cr
Additional arguments passed to the respective \link[mlr3]{DataBackend} method.}

\item{polygons}{\verb{[logical]}\cr
Whether to convert to polygons instead of points.}

\item{quiet}{\verb{[logical]}\cr
Whether to suppress possible console output.}
}
\value{
\link{DataBackend}.
}
\description{
S3 dispatch methods for converting from and to various spatial \link[mlr3:DataBackend]{mlr3::DataBackend}s.
}
\seealso{
DataBackendSf DataBackendStars
}
