% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataBackendRaster.R
\name{as_data_backend.stars}
\alias{as_data_backend.stars}
\alias{as_data_backend.SpatRaster}
\alias{as_data_backend.RasterBrick}
\alias{as_data_backend.RasterStack}
\title{Coerce to spatial DataBackend}
\usage{
\method{as_data_backend}{stars}(data, primary_key = NULL, ...)

\method{as_data_backend}{SpatRaster}(data, primary_key = NULL, ...)

\method{as_data_backend}{RasterBrick}(data, primary_key = NULL, ...)

\method{as_data_backend}{RasterStack}(data, primary_key = NULL, ...)
}
\arguments{
\item{data}{(\link[terra:SpatRaster-class]{terra::SpatRaster})\cr
The input \link[terra:SpatRaster-class]{terra::SpatRaster}.}

\item{primary_key}{(\code{character(1)} | \code{integer()})\cr
Name of the primary key column, or integer vector of row ids.}

\item{...}{(\code{any})\cr
Not used.}
}
\value{
\link{DataBackend}.
}
\description{
Wraps a \link{DataBackend} around spatial objects.
Currently these S3 methods are only alternative ways for writing \code{DataBackendRaster$new()} or \code{DataBackendVector$new()}.
They do not support coercing from other backends yet.
}
