% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataBackendSpatial.R
\name{as_data_backend.stars}
\alias{as_data_backend.stars}
\title{Create a Data Backend}
\usage{
\method{as_data_backend}{stars}(data, primary_key = NULL, ...)
}
\arguments{
\item{data}{(\code{SpatRaster})\cr
A \code{SpatRaster} object to create a \link{DataBackend} from.}

\item{primary_key}{(\code{character(1)} | \code{integer()})\cr
Name of the primary key column, or integer vector of row ids.}

\item{...}{(\code{any})\cr
Not used.}
}
\value{
\link{DataBackend}.
}
\description{
Wraps a \link{DataBackend} around \code{SpatRaster} objects.
}
