% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{block_size}
\alias{block_size}
\title{Split Raster Into Chunks}
\usage{
block_size(raster, chunksize)
}
\arguments{
\item{raster}{(\link[terra:SpatRaster-class]{terra::SpatRaster})\cr
Raster to be split into chunks.}

\item{chunksize}{Raster chunk size in megabyte.}
}
\description{
Splits raster into chunks.
}
