% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataBackendSpatRaster.R
\name{DataBackendSpatRaster}
\alias{DataBackendSpatRaster}
\title{DataBackend for SpatRaster}
\description{
A \link[mlr3:DataBackend]{mlr3::DataBackend} for \code{SpatRaster} (\CRANpkg{terra}).
}
\section{Read mode}{

\itemize{
\item Block mode reads complete rows of the raster file and subsets the requested
cells. Faster than cell mode if we iterate the whole raster file.
\item Cell mode reads individual cells. Faster than block mode if only a few
cells are sampled.
}

Block mode is activated if \verb{$data(rows)} is called with a increasing integer
sequence e.g. \code{200:300}.
}

\section{Super class}{
\code{\link[mlr3:DataBackend]{mlr3::DataBackend}} -> \code{DataBackendSpatRaster}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{rownames}}{(\code{integer()})\cr
Returns vector of all distinct row identifiers, i.e. the contents of the primary key column.}

\item{\code{colnames}}{(\code{character()})\cr
Returns vector of all column names.}

\item{\code{nrow}}{(\code{integer(1)})\cr
Number of rows (observations).}

\item{\code{ncol}}{(\code{integer(1)})\cr
Number of columns (variables).}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{DataBackendSpatRaster$new()}}
\item \href{#method-data}{\code{DataBackendSpatRaster$data()}}
\item \href{#method-finalize}{\code{DataBackendSpatRaster$finalize()}}
\item \href{#method-head}{\code{DataBackendSpatRaster$head()}}
\item \href{#method-distinct}{\code{DataBackendSpatRaster$distinct()}}
\item \href{#method-missings}{\code{DataBackendSpatRaster$missings()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="mlr3" data-topic="DataBackend" data-id="format">}\href{../../mlr3/html/DataBackend.html#method-format}{\code{mlr3::DataBackend$format()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3" data-topic="DataBackend" data-id="print">}\href{../../mlr3/html/DataBackend.html#method-print}{\code{mlr3::DataBackend$print()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Creates a backend for a \code{SpatRaster}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataBackendSpatRaster$new(data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{(\code{SpatRaster})\cr
A raster object.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-data"></a>}}
\if{latex}{\out{\hypertarget{method-data}{}}}
\subsection{Method \code{data()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataBackendSpatRaster$data(rows, cols, data_format)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{rows}}{\code{integer()}\cr
Row indices. Row indices start with 1 in the upper left corner in the
raster, increase from left to right and then from top to bottom. The last
cell is in the bottom right corner and the row index equals the number of
cells in the raster.}

\item{\code{cols}}{\code{character()}\cr
Column names.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-finalize"></a>}}
\if{latex}{\out{\hypertarget{method-finalize}{}}}
\subsection{Method \code{finalize()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataBackendSpatRaster$finalize()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-head"></a>}}
\if{latex}{\out{\hypertarget{method-head}{}}}
\subsection{Method \code{head()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataBackendSpatRaster$head(n = 6L)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-distinct"></a>}}
\if{latex}{\out{\hypertarget{method-distinct}{}}}
\subsection{Method \code{distinct()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataBackendSpatRaster$distinct(rows, cols, na_rm = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{rows}}{\code{integer()}\cr
Row indices. Row indices start with 1 in the upper left corner in the
raster, increase from left to right and then from top to bottom. The last
cell is in the bottom right corner and the row index equals the number of
cells in the raster.}

\item{\code{cols}}{\code{character()}\cr
Column names.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-missings"></a>}}
\if{latex}{\out{\hypertarget{method-missings}{}}}
\subsection{Method \code{missings()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataBackendSpatRaster$missings(row, cols)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{cols}}{\code{character()}\cr
Column names.}
}
\if{html}{\out{</div>}}
}
}
}
