% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataBackendVector.R
\name{DataBackendVector}
\alias{DataBackendVector}
\title{DataBackend for Vector Objects}
\description{
\link[mlr3:DataBackend]{mlr3::DataBackend} for \link[sf:sf]{sf::sf} vector objects.
}
\section{Super classes}{
\code{\link[mlr3:DataBackend]{mlr3::DataBackend}} -> \code{\link[mlr3:DataBackendDataTable]{mlr3::DataBackendDataTable}} -> \code{DataBackendVector}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{sfc}}{(\link[sf:sfc]{sf::sfc})\cr
Returns the sfc object.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DataBackendVector-new}{\code{DataBackendVector$new()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="DataBackend" data-id="format"><a href='../../mlr3/html/DataBackend.html#method-DataBackend-format'><code>mlr3::DataBackend$format()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="DataBackend" data-id="print"><a href='../../mlr3/html/DataBackend.html#method-DataBackend-print'><code>mlr3::DataBackend$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="DataBackendDataTable" data-id="data"><a href='../../mlr3/html/DataBackendDataTable.html#method-DataBackendDataTable-data'><code>mlr3::DataBackendDataTable$data()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="DataBackendDataTable" data-id="distinct"><a href='../../mlr3/html/DataBackendDataTable.html#method-DataBackendDataTable-distinct'><code>mlr3::DataBackendDataTable$distinct()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="DataBackendDataTable" data-id="head"><a href='../../mlr3/html/DataBackendDataTable.html#method-DataBackendDataTable-head'><code>mlr3::DataBackendDataTable$head()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="DataBackendDataTable" data-id="missings"><a href='../../mlr3/html/DataBackendDataTable.html#method-DataBackendDataTable-missings'><code>mlr3::DataBackendDataTable$missings()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataBackendVector-new"></a>}}
\if{latex}{\out{\hypertarget{method-DataBackendVector-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this \link[R6:R6Class]{R6} class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataBackendVector$new(data, primary_key)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{(\code{sf})\cr
A raster object.}

\item{\code{primary_key}}{(\code{character(1)} | \code{integer()})\cr
Name of the primary key column, or integer vector of row ids.}
}
\if{html}{\out{</div>}}
}
}
}
