% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{demo_stack}
\alias{demo_stack}
\title{Generate Demo Raster Stack}
\usage{
demo_stack(size = 500, layers = 5)
}
\arguments{
\item{size}{(\code{integer(1)})\cr
Size of raster stack in megabyte.}

\item{layer}{(\code{integer(1)})\cr
Number of layers.}
}
\description{
Generates a square demo \link[terra:SpatRaster-class]{terra::SpatRaster} stack.
}
