% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{factor_layer}
\alias{factor_layer}
\title{Categorical Raster Layer}
\usage{
factor_layer(id, levels, in_memory = FALSE)
}
\arguments{
\item{id}{\code{character(1)}\cr
Layer id.}

\item{levels}{\code{character()}\cr
Categories of the layer.}

\item{in_memory}{\code{logical(1)}\cr
If \code{FALSE} (default), layer is written to disk.}
}
\description{
Creates a categorical raster layer when passed to \code{create_stack()}.
}
