% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{generate_stack}
\alias{generate_stack}
\title{Generate Raster Stack}
\usage{
generate_stack(layers, layer_size = NULL, dimension = NULL)
}
\arguments{
\item{layers}{(List of \code{\link[=numeric_layer]{numeric_layer()}} and \code{\link[=factor_layer]{factor_layer()}})\cr
List of layers.}

\item{layer_size}{(\code{numeric(1)})\cr
Size of a single layer in megabytes.}

\item{dimension}{(\code{integer(1)})\cr
Dimension of the squared layers.

\code{layer_size} and \code{dimension} are mutually exclusive.}
}
\value{
\link[terra:SpatRaster-class]{terra::SpatRaster}
}
\description{
Generates a raster stack.
}
\keyword{internal}
