% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{mlr3spatial-package}
\alias{mlr3spatial}
\alias{mlr3spatial-package}
\title{mlr3spatial: Parallel spatial prediction for 'mlr3'}
\description{
Parallel raster prediction for mlr3.
}
\references{
\cite{mlr3spatial}{pkg::citation}
}
\author{
\strong{Maintainer}: Marc Becker \email{marcbecker@posteo.de}

}
