% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{numeric_layer}
\alias{numeric_layer}
\title{Numeric Layer Generator}
\usage{
numeric_layer(id, in_memory = FALSE)

numeric_layer(id, in_memory = FALSE)
}
\arguments{
\item{id}{\code{character(1)}\cr
Layer id.}

\item{in_memory}{\code{logical(1)}\cr
If \code{FALSE} (default), layer is written to disk.}
}
\value{
Named \code{list()}
}
\description{
Generates a numeric layer when passed to \code{\link[=generate_stack]{generate_stack()}}.

Creates a numeric raster layer when passed to \code{create_stack()}.
}
\keyword{internal}
